refactor : #8 DogService Interface 추가
