+++ src/main/java/com/ddang/dog/service/DogService.java
@@ -1,130 +1,22 @@
 package com.ddang.dog.service;
-//깃허브 데스크탑
-//테스트
 
-import com.ddang.dog.entity.Dog;
-import com.ddang.dog.entity.MemberDog;
-import com.ddang.dog.repository.DogRepository;
-import com.ddang.dog.repository.MemberDogRepository;
 import com.ddang.dog.service.request.CreateDogServiceRequest;
 import com.ddang.dog.service.request.UpdateDogServiceRequest;
 import com.ddang.dog.service.response.DogResponse;
-import com.ddang.family.entity.Family;
-import com.ddang.family.repository.FamilyRepository;
-import com.ddang.global.exception.BadRequestException;
-import com.ddang.global.exception.ErrorCode;
-import com.ddang.global.service.S3Service;
 import com.ddang.member.entity.Member;
-import com.ddang.member.repository.MemberRepository;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
 import java.util.List;
 
+public interface DogService {
+    DogResponse createDog(CreateDogServiceRequest request, Member member, MultipartFile profileImgFile) throws IOException;
 
-@Slf4j
-@Service
-@RequiredArgsConstructor
-@Transactional
-public class DogService {
+    DogResponse getDogByDogId(Long dogId);
 
-    private final DogRepository dogRepository;
-    private final MemberRepository memberRepository;
-    private final MemberDogRepository memberDogRepository;
-    private final FamilyRepository familyRepository;
-    private final S3Service s3Service;
+    DogResponse updateDog(UpdateDogServiceRequest request,Long dogId, Long memberId, MultipartFile profileImgFile) throws IOException;
 
-    private final static String DOG_PROFILE_DIR = "dog";
+    void deleteDog(Long dogId, Long memberId);
 
-    public DogResponse createDog(CreateDogServiceRequest request, Member member, MultipartFile profileImgFile) throws IOException {
-        String profileImg = s3Service.upload(profileImgFile, DOG_PROFILE_DIR);
-        createFamilyIfNotExists(member);
-        Dog dog = request.toEntity(profileImg, member.getFamily());
-        List<MemberDog> memberDog = assignDogToFamilyMembers(member, dog);
-
-        dogRepository.save(dog);
-        memberDogRepository.saveAll(memberDog);
-
-        return DogResponse.from(dog);
-    }
-
-    public DogResponse getDogByDogId(Long dogId) {
-        Dog dog = findDogByIdOrThrowException(dogId);
-
-        return DogResponse.from(dog);
-    }
-
-    public void updateDog(UpdateDogServiceRequest request, Long memberId, MultipartFile profileImgFile) throws IOException {
-
-        MemberDog memberDog = memberDogRepository.findByDogIdAndMemberId(request.dogId(), memberId)
-                .orElseThrow(() -> new BadRequestException(ErrorCode.MEMBER_NOT_HAVE_DOG));
-
-        Dog dog = memberDog.getDog();
-        String profileImg = getProfileImgUrlOrElseGetNull(profileImgFile);
-
-        dog.update(request, profileImg);
-    }
-
-    public void deleteDog(Long dogId, Long memberId) {
-        memberDogRepository.findByDogIdAndMemberId(dogId, memberId)
-                .orElseThrow(() -> new BadRequestException(ErrorCode.MEMBER_NOT_HAVE_DOG));
-
-        memberDogRepository.softDeleteByDogId(dogId);
-
-        dogRepository.softDeleteById(dogId);
-        // TODO 산책 내역 삭제하기
-    }
-
-    public List<DogResponse> getDogsByMember(Member member) {
-        List<MemberDog> memberDogs = memberDogRepository.findAllByMember(member);
-
-        return memberDogs.stream()
-                .map(memberDog -> DogResponse.from(memberDog.getDog()))
-                .toList();
-    }
-
-
-    private Dog findDogByIdOrThrowException(Long id) {
-        return dogRepository.findActiveById(id)
-                .orElseThrow(() -> {
-                    log.warn(">>>> {} : {} <<<<", id, ErrorCode.DOG_NOT_FOUND);
-                    return new BadRequestException(ErrorCode.DOG_NOT_FOUND);
-                });
-    }
-
-    private String getProfileImgUrlOrElseGetNull(MultipartFile profileImgFile) throws IOException {
-        if (profileImgFile != null) {
-            return s3Service.upload(profileImgFile, DOG_PROFILE_DIR);
-        }
-
-        return null;
-    }
-
-    private void createFamilyIfNotExists(Member member){
-        if(member.hasNoFamily()) {
-            Family family = Family.create();
-            family = familyRepository.save(family);
-
-            member.updateFamily(family);
-        }
-    }
-
-    private List<MemberDog> assignDogToFamilyMembers(Member member, Dog dog){
-
-        List<Member> members = memberRepository.findAllByFamily(member.getFamily());
-        return members.stream()
-                .map(familyMember -> MemberDog.builder()
-                        .member(familyMember)
-                        .dog(dog)
-                        .build())
-                .toList();
-    }
-
-
+    List<DogResponse> getDogsByMember(Member member);
 }
-
-
